# Use LibC.malloc instead of GC.malloc for LibZ allocations