--- a/src/compress/deflate/writer.cr
+++ b/src/compress/deflate/writer.cr
@@ -20,12 +20,16 @@ class Compress::Deflate::Writer < IO
 
     @buf = uninitialized UInt8[8192] # output buffer used by zlib
     @stream = LibZ::ZStream.new
-    @stream.zalloc = LibZ::AllocFunc.new { |opaque, items, size| GC.malloc(items * size) }
-    @stream.zfree = LibZ::FreeFunc.new { |opaque, address| GC.free(address) }
     @closed = false
     ret = LibZ.deflateInit2(pointerof(@stream), level, LibZ::Z_DEFLATED, -LibZ::MAX_BITS, LibZ::DEF_MEM_LEVEL,
       strategy.value, LibZ.zlibVersion, sizeof(LibZ::ZStream))
     raise Compress::Deflate::Error.new(ret, @stream) unless ret.ok?
+  end
+
+  def finalize
+    return if closed?
+
+    close rescue nil
   end
 
   # Creates a new writer for the given *io*, yields it to the given block,