Follow-up on #12451.

This makes the memory management os LibZ manual for the stdlib. So any internal allocations are made with the system allocator, leaving less memory being managed by the GC.

Public interface is mostly unchanged, except for the fact that `Compress::Deflate::Reader` and `Compress::Deflate::Writer` are now closed on finalize.

Note: We currently don't close IOs on finalize. I think we should. `IO::Buffered`, for example, could leave unwritten data on its buffer.