@oprypin 2022-09-07
Have you benchmarked this? Adding a `finalize` method is heavy and could in some cases outweigh the benefits.

> Note: We currently don't close IOs on finalize. I think we should

I don't think I agree... Code that accidentally mostly works is worse than never working.