@lbguilherme 2022-09-08
> > Note: We currently don't close IOs on finalize. I think we should
>
> I don't think I agree... Code that accidentally mostly works is worse than never working.

We definitely shouldn't! Closing a buffered IO on finalize means performing IO inside a GC cycle, which in turn could mean that a Fiber swap can happen during the `finalize` execution and that it could take a long time do return and block the GC itself from continuing. Let's not create this problem. I've also looked up and neither Java nor Go try to flush streams on finalize. The finalize method should concern about releasing memory and OS resources, nothing more. Thanks for the call out.

I pushed a commit ensuring that the `finalize` method doesn't do anything other than freeing memory.

> Have you benchmarked this? Adding a finalize method is heavy and could in some cases outweigh the benefits.

I did a benchmark now and the result is quite surprising.

```cr
require "benchmark"
require "compress/deflate"

io = IO::Memory.new

Benchmark.ips do |x|
  5.times do |i|
    x.report("leaking many writers (#{i})") do
      GC.collect
      10_000.times do
        Compress::Deflate::Writer.new io
      end
      GC.collect
    end
  end
end
```

Without this PR:

```
leaking many writers (0) 108.41m (  9.22s ) (± 0.00%)  2.58GB/op   1.04× slower
leaking many writers (1) 104.64m (  9.56s ) (± 0.00%)  2.58GB/op   1.08× slower
leaking many writers (2) 103.05m (  9.70s ) (± 0.00%)  2.58GB/op   1.09× slower
leaking many writers (3) 112.52m (  8.89s ) (± 0.00%)  2.58GB/op        fastest
leaking many writers (4) 109.79m (  9.11s ) (± 0.00%)  2.58GB/op   1.02× slower
```

With this PR:
```
leaking many writers (0) 866.56m (  1.15s ) (± 2.37%)  80.9MB/op   3.44× slower
leaking many writers (1)   1.30  (770.66ms) (± 3.69%)  80.9MB/op   2.30× slower
leaking many writers (2)   2.09  (478.48ms) (±33.67%)  80.9MB/op   1.43× slower
leaking many writers (3)   2.64  (379.02ms) (±48.81%)  80.9MB/op   1.13× slower
leaking many writers (4)   2.98  (335.71ms) (±31.39%)  80.9MB/op        fastest
```

This is a 10x to 20x speed improvement. Doesn't make much sense to me.

I also ran this code to ensure that there is no memory leak:

```cr
require "benchmark"
require "compress/deflate"

io = IO::Memory.new

while true
  Compress::Deflate::Writer.new io
end
```

And memory does not grow over time. Everything is being properly collected.