@HertzDevil 2022-09-17
This also crashes the interpreter. Is this related to #12456?