@hugopl 2022-09-28
I think this is a good MR and probably solves the same problem of https://github.com/crystal-lang/crystal/issues/10435 but for LibZ.