@hugopl 2022-09-28
Sorry, my bad, MR is PR, Gitlab calls "merge request", as I use Gitlab at work I'm used to call PR as MR. :sweat_smile: 

MR: merge request
PR: pull request