@straight-shoota 2022-09-28
> This also crashes the interpreter. Is this related to #12456?

The interpreter failure is probably just an appearance of an existing error (#12528) and unrelated to this PR. I've triggered a re-run to check that.