@straight-shoota 2022-09-28
I ran the same benchmark because the numbers seemed a bit too good to be true. I see about 2x speed improvement. But I'm also not entirely sure how much the benchmark can actually be trusted.

Without this PR:
```
                               user     system      total        real
leaking many writers (0)   0.321653   0.000000   0.321653 (  0.321714)
leaking many writers (1)   0.313898   0.000000   0.313898 (  0.313900)
leaking many writers (2)   0.316821   0.000000   0.316821 (  0.316824)
leaking many writers (3)   0.318118   0.000000   0.318118 (  0.318120)
leaking many writers (4)   0.320555   0.000000   0.320555 (  0.320556)
```
With this PR:
```
                               user     system      total        real
leaking many writers (0)   0.148214   0.000000   0.148214 (  0.148260)
leaking many writers (1)   0.000000   0.139552   0.139552 (  0.139578)
leaking many writers (2)   0.000000   0.145247   0.145247 (  0.145331)
leaking many writers (3)   0.033282   0.021551   0.054833 (  0.054835)
leaking many writers (4)   0.134505   0.011903   0.146408 (  0.146411)
```