@beta-ziliani 2024-07-19
I'm not sure what's the right path here, but what is certain is that if we want GC to take care of the allocation, then we must call `GC.malloc_atomic` instead.