@ysbaddaden 2024-07-19
The above benchmark seems faster (400ms -> 150s) but the benchmark from #11614 is twice slower (2.6s -> 4.9s) on Ubuntu 20.04 :shrug:

I tried `GC.malloc_atomic` but it segfaults.